refactor: extract host-agnostic data app viz builder pieces
Lift AppPreview out of AppGenerate, split the restore confirm modal and
version provenance line into standalone components, and pull the test-
context state machine out of DataAppVizTestPanel so a second authoring
surface can reuse them. Adds the chart_type_builder creation experience
to the common union. No behavior change.

# Conflicts:
#	packages/frontend/src/pages/AppGenerate.tsx
