+++ packages/common/src/ee/apps/types.ts
@@ -112,6 +112,7 @@ export type DataAppTemplate = (typeof DATA_APP_TEMPLATES)[number];
 export const DATA_APP_CREATION_EXPERIENCES = [
     'app_builder',
     'explorer_chart_config',
+    'chart_type_builder',
 ] as const;
 export type DataAppCreationExperience =
     (typeof DATA_APP_CREATION_EXPERIENCES)[number];
